qrcode-restaurante/backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SocketIO==5.3.6
PyJWT==2.8.0
bcrypt==4.1.3
psycopg2-binary==2.9.9; python_version < "3.13"
pg8000==1.31.2
python-dotenv==1.0.1
qrcode==7.4.2
eventlet==0.35.2; python_version < "3.13"
requests==2.31.0
gunicorn==22.0.0
